refactor(Chat): change filenames of chat system
New filenames for chat are as following
- chatRoom -> chatConversation
- chatInput -> chatSendMessage
- chatOffer -> chatHeader
- chatList -> chatRooms
