+++ src/routes/chat/ChatHeader/index.js
@@ -1,0 +1,1 @@
+export { default } from './ChatDefaultHeader';
