--- src/routes/chat/ChatOffer/index.js
@@ -1,1 +1,0 @@
-export { default } from './ChatOffer';
