+++ src/routes/chat/ChatRooms/index.js
@@ -1,0 +1,1 @@
+export { default } from './ChatRooms';
